Added routes for pages
