+++ src/app.css
@@ -181,7 +181,7 @@
 	width: 100%;
 	height: 100%;
 
-	padding: 2rem;
+	padding-top: 1.75%;
 
 	overflow-y: auto;
 }
