--- src/routes/chat/+page.svelte
@@ -1,1 +1,0 @@
-chat page
