+++ src/routes/legal/privacy/+page.svelte
@@ -2,9 +2,11 @@
 	import { m } from '$lib/paraglide/messages';
 </script>
 
-<h1 class="terms">
-	{m.privacy_policy()}
-</h1>
+<div class="glass w-full h-full p-6">
+	<h1 class="terms">
+		{m.privacy_policy()}
+	</h1>
+</div>
 
 <style>
 	.terms {
