+++ src/routes/profile/+page.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+</script>
+
+<div class="glass w-full h-full p-6">
+	profile page
+</div>
+
+<style>
+</style>
