+++ src/routes/settings/+page.svelte
@@ -1,0 +1,24 @@
+<script lang="ts">
+	import { useSession, signOut } from '$lib/auth-client';
+	import { Button } from '$lib/components/ui/button';
+
+	const session = useSession();
+</script>
+
+<div class="glass w-full h-full p-6">
+	<div>
+		{#if $session.data}
+			{#each Object.entries($session.data.user) as [key, value] (key)}
+				<p><strong>{key}:</strong> {value}</p>
+			{/each}
+			<Button
+				class="glass"
+				onclick={async () => {
+					await signOut();
+				}}>Logout</Button
+			>
+		{:else}
+			<p>Not logged in</p>
+		{/if}
+	</div>
+</div>
